requirements.txt
gradio==4.10.0
python-dotenv==1.0.0
jupyter==1.0.0
openai==1.5.0
pandas==2.1.4
matplotlib==3.8.2
requests~=2.31.0
pillow==10.1.0
inquirer==3.1.4
rich==13.7.0
appdirs==1.4.4
packaging==23.2